requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
accelerate=1.5.2=pypi_0
aiofiles=23.2.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
audioread=3.0.1=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
decorator=5.2.1=pypi_0
exceptiongroup=1.2.2=pypi_0
fastapi=0.115.12=pypi_0
fastdtw=0.3.4=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.56.0=pypi_0
fsspec=2025.3.0=pypi_0
gammatone=1.0=pypi_0
gradio=5.23.0=pypi_0
gradio-client=1.8.0=pypi_0
groovy=0.1.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.29.3=pypi_0
idna=3.10=pypi_0
importlib-resources=6.5.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
libffi=3.4.4=hd77b12b_1
librosa=0.11.0=pypi_0
llvmlite=0.44.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=pypi_0
mdurl=0.1.2=pypi_0
mock=5.2.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
networkx=3.4.2=pypi_0
nose=1.3.7=pypi_0
numba=0.61.0=pypi_0
openssl=3.0.16=h3f729d1_0
orjson=3.10.16=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
peft=0.14.0=pypi_0
pesq=0.0.5=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py310haa95532_0
platformdirs=4.3.7=pypi_0
pooch=1.8.2=pypi_0
psutil=7.0.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.3=pypi_0
pysepm=0.1=pypi_0
pystoi=0.4.1=pypi_0
python=3.10.14=he1021f5_1
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
ruff=0.11.2=pypi_0
safehttpx=0.1.6=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=75.8.0=py310haa95532_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soundfile=0.13.1=pypi_0
soxr=0.5.0.post1=pypi_0
sqlite=3.45.3=h2bbff1b_0
srmrpy=1.0=pypi_0
starlette=0.46.1=pypi_0
sympy=1.13.1=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.21.1=pypi_0
tomlkit=0.13.2=pypi_0
torch=1.12.1=pypi_0
torchaudio=0.12.1=pypi_0
torchlibrosa=0.1.0=pypi_0
torchvision=0.13.1=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.47.1=pypi_0
typer=0.15.2=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
websockets=15.0.1=pypi_0
wheel=0.45.1=py310haa95532_0
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
